Support specifying the name of the default pool to use (#16)
Currently, it is only possible to specify the pool name of large actions. With this PR, you can also specify a custom pool name for all actions to be executed on. This is helpful if actions are not to be executed on the RBE service's default pool (used when no pool is specified).